+++ R/Spectra.R
@@ -353,6 +353,10 @@ NULL
 #'   provided m/z range. See examples for details on selecting spectra with
 #'   a precursor m/z for a target m/z accepting a small difference in *ppm*.
 #'
+#' - `filterPrecursorMzValues`: retains spectra with precursor m/z matching any
+#'   of the provided m/z values (given `ppm` and `tolerance`). Spectra with
+#'   missing precursor m/z value (e.g. MS1 spectra) are dropped.
+#'
 #' - `filterPrecursorCharge`: retains spectra with the defined precursor
 #'   charge(s).
 #'
@@ -702,8 +706,8 @@ NULL
 #' @param mz For `filterIsolationWindow`: `numeric(1)` with the m/z value to
 #'     filter the object. For `filterPrecursorMz` and `filterMzRange`:
 #'     `numeric(2)` defining the lower and upper m/z boundary.
-#'     For `filterMzValues`: `numeric` with the m/z values to match peaks
-#'     against.
+#'     For `filterMzValues` and `filterPrecursorMzValues`: `numeric` with the
+#'     m/z values to match peaks or precursor m/z against.
 #'
 #' @param z For `filterPrecursorCharge`: `integer()` with the precursor charges
 #'     to be used as filter.
@@ -1940,6 +1944,18 @@ setMethod("filterPrecursorMz", "Spectra",
                   object@processing,
                   "Filter: select spectra with a precursor m/z within [",
                   paste0(mz, collapse = ", "), "]")
+              object
+          })
+
+#' @rdname Spectra
+setMethod("filterPrecursorMzValues", "Spectra",
+          function(object, mz = numeric(), ppm = 20, tolerance = 0) {
+              object@backend <- filterPrecursorMzValues(
+                  object@backend, mz, ppm = ppm, tolerance = tolerance)
+              object@processing <- .logging(
+                  object@processing,
+                  "Filter: select spectra with precursor m/z matching ",
+                  paste0(mz, collapse = ", "), "")
               object
           })